requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
python-dotenv
joblib
scikit-learn
jsonschema
numpy
